requirements.txt
redis==5.2.1
Pillow==10.1.0
requests==2.31.0
aiohttp==3.9.1
pywebview==5.4
